+++ src/gpu/intel/softmax/reusable.hpp
@@ -190,7 +190,7 @@ struct reusable_fwd_t : public primitive_t {
             rt_conf.softmax_axis_size = src_mdw.dims()[desc()->softmax_axis];
             for (const auto &block : layout) {
                 if (block.dim_idx == into<dim_idx_t>(desc()->softmax_axis)) {
-                    rt_conf.softmax_axis_stride = block.stride;
+                    rt_conf.softmax_axis_stride = int64_t(block.stride);
                     break;
                 }
             }
